Saturator working with w/d

+++ C Programs/Final/src/global.c
@@ -1,1 +1,2 @@
-float lookUp[300];
+#include <stdint.h>
+uint16_t lookUp[300];
