+++ C Programs/Final/src/global.h
@@ -1,1 +1,2 @@
-extern float lookUp[300];
+#include <stdint.h>
+extern uint16_t lookUp[300];
